moved algorithm into new file
